stage

+++ back_end/restservice/src/main/java/com/goj/restservice/controller/ContestController.java
@@ -1,6 +1,7 @@
 package com.goj.restservice.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
@@ -27,7 +28,10 @@ import com.goj.restservice.exception.CustomException;
 import com.goj.restservice.form.ContestForm;
 import com.goj.restservice.projection.ContestDetail;
 import com.goj.restservice.projection.ContestSummary;
+import com.goj.restservice.projection.SubmissionSummary;
+import com.goj.restservice.repository.ContestRepository;
 import com.goj.restservice.repository.ContestUserRepository;
+import com.goj.restservice.repository.SubmissionRepository;
 import com.goj.restservice.service.ContestService;
 import com.goj.restservice.util.Util;
 
@@ -39,17 +43,24 @@ public class ContestController {
     private ContestService contestService;
 
     @Autowired
+    private ContestRepository contestRepository;
+
+    @Autowired
     Util util;
 
     @Autowired
-    ContestUserRepository contestUserRepository;
+    private ContestUserRepository contestUserRepository;
+
+    @Autowired
+    private SubmissionRepository submissionRepository;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public void create(@Valid @RequestBody ContestForm contestForm, HttpServletRequest request,
             HttpServletResponse response, @AuthenticationPrincipal User user) {
-        Contest newContest = new Contest(contestForm.getTitle(), contestForm.getDescription(),
-                contestForm.getStartTime(), contestForm.getEndTime(), contestForm.getPassword());
+        Contest newContest = new Contest();
+        newContest.update(contestForm.getTitle(), contestForm.getDescription(), contestForm.getStartTime(),
+                contestForm.getEndTime(), contestForm.getPassword());
 
         newContest.setCreateUser(user);
         Contest createdContest = contestService.create(newContest);
@@ -61,8 +72,31 @@ public class ContestController {
     @GetMapping
     public @ResponseBody Iterable<ContestSummary> readAll(
             @RequestParam(value = "page", defaultValue = "1") @Min(value = 1, message = "page must be greater than or equal to 1") int page,
-            @RequestParam(value = "per_page", defaultValue = "10000") @Min(value = 1, message = "per_page must be greater than or equal to 1") int per_page) {
-        return contestService.readAll(page - 1, per_page);
+            @RequestParam(value = "per_page", defaultValue = "10000") @Min(value = 1, message = "per_page must be greater than or equal to 1") int per_page,
+            @AuthenticationPrincipal User user) {
+        if (user == null)
+            return contestService.readAll(page - 1, per_page);
+        else
+            return contestRepository.findAllContestSummaryByUserId(user.getUserId(),
+                    PageRequest.of(page - 1, per_page));
+    }
+
+    @GetMapping("/{contestId}/submissions")
+    public @ResponseBody Iterable<SubmissionSummary> readAllSubmissions(@PathVariable("contestId") Long contestId,
+            @RequestParam(value = "page", defaultValue = "1") @Min(value = 1, message = "page must be greater than or equal to 1") int page,
+            @RequestParam(value = "per_page", defaultValue = "10000") @Min(value = 1, message = "per_page must be greater than or equal to 1") int per_page,
+            @AuthenticationPrincipal User user) {
+
+        ContestDetail contestDetail = contestService.readOne(contestId);
+        util.checkResourceFound(contestDetail);
+
+        if (contestDetail.getCreateUserUsername().equals(user.getUsername()) || user.getRoles().contains("ROLE_GMH")
+                || contestUserRepository.existsByContestIdAndUserId(contestId, user.getUserId())) {
+            return submissionRepository.findAllSubmissionSummaryByContestId(contestId,
+                    PageRequest.of(page - 1, per_page));
+        } else {
+            throw new CustomException("User doesn't join the contest.", HttpStatus.BAD_REQUEST);
+        }
     }
 
     @GetMapping("/{contestId}")
@@ -71,31 +105,43 @@ public class ContestController {
         ContestDetail contestDetail = contestService.readOne(contestId);
         util.checkResourceFound(contestDetail);
 
-        if (contestDetail.getCreateUserUsername().equals(user.getUsername()) || user.getRoles().contains("ROLE_GMH"))
+        if (contestDetail.getCreateUserUsername().equals(user.getUsername()) || user.getRoles().contains("ROLE_GMH")
+                || contestUserRepository.existsByContestIdAndUserId(contestId, user.getUserId()))
             return contestDetail;
-
-        if (!contestUserRepository.existsByContestIdAndUserId(contestId, user.getUserId()))
+        else
             throw new CustomException("User doesn't join the contest.", HttpStatus.BAD_REQUEST);
 
-        return contestDetail;
     }
 
     @PutMapping("/{contestId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void update(@PathVariable("contestId") Long contestId, @Valid @RequestBody ContestForm contestForm,
             HttpServletRequest request, HttpServletResponse response, @AuthenticationPrincipal User user) {
-        Contest newContest = new Contest(contestForm.getTitle(), contestForm.getDescription(),
-                contestForm.getStartTime(), contestForm.getEndTime(), contestForm.getPassword());
 
-        newContest.setCreateUser(user);
-        newContest.setContestId(contestId);
-        contestService.update(newContest);
+        Contest contest = new Contest();
+        if (contestRepository.existsById(contestId)) {
+            contest = contestRepository.findById(contestId).get();
+            if (contest.getCreateUser().getUserId() != user.getUserId() && !user.getRoles().contains("ROLE_GMH"))
+                throw new CustomException("Method not allow", HttpStatus.BAD_REQUEST);
+        }
+        contest.update(contestForm.getTitle(), contestForm.getDescription(), contestForm.getStartTime(),
+                contestForm.getEndTime(), contestForm.getPassword());
+
+        contest.setContestId(contestId);
+        contestService.update(contest);
     }
 
     @DeleteMapping("/{contestId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable("contestId") Long contestId, HttpServletRequest request,
-            HttpServletResponse response) {
-        contestService.delete(contestId);
+            HttpServletResponse response, @AuthenticationPrincipal User user) {
+
+        if (contestRepository.existsById(contestId)) {
+            Contest contest = contestRepository.findById(contestId).get();
+            if (contest.getCreateUser().getUserId() != user.getUserId() && !user.getRoles().contains("ROLE_GMH"))
+                throw new CustomException("Method not allow", HttpStatus.BAD_REQUEST);
+
+            contestService.delete(contestId);
+        }
     }
 }
