+++ back_end/restservice/src/main/java/com/goj/restservice/controller/ProblemController.java
@@ -1,7 +1,9 @@
 package com.goj.restservice.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -21,8 +23,11 @@ import javax.validation.Valid;
 import javax.validation.constraints.Min;
 
 import com.goj.restservice.entity.Problem;
+import com.goj.restservice.entity.User;
+import com.goj.restservice.exception.CustomException;
 import com.goj.restservice.form.ProblemForm;
 import com.goj.restservice.projection.ProblemSummary;
+import com.goj.restservice.repository.ProblemRepository;
 import com.goj.restservice.service.ProblemService;
 
 import com.goj.restservice.util.Util;
@@ -35,16 +40,23 @@ public class ProblemController {
     private ProblemService problemService;
 
     @Autowired
+    private ProblemRepository problemRepository;
+
+    @Autowired
     Util util;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public void create(@Valid @RequestBody ProblemForm problemForm, HttpServletRequest request,
-            HttpServletResponse response) {
-        Problem newProblem = new Problem(problemForm.getTitle(), problemForm.getSource(), problemForm.getDescription(),
+            HttpServletResponse response, @AuthenticationPrincipal User user) {
+
+        Problem newProblem = new Problem();
+        newProblem.update(problemForm.getTitle(), problemForm.getSource(), problemForm.getDescription(),
                 problemForm.getInput(), problemForm.getOutput(), problemForm.getSampleInput(),
                 problemForm.getSampleOutput(), problemForm.getHint(), problemForm.getTimeLimit(),
                 problemForm.getMemoryLimit());
+        newProblem.setCreateUser(user);
+
         Problem createProblem = problemService.create(newProblem);
         response.setHeader("Location",
                 request.getRequestURL().append("/").append(createProblem.getProblemId()).toString());
@@ -54,8 +66,14 @@ public class ProblemController {
     @GetMapping
     public @ResponseBody Iterable<ProblemSummary> readAll(
             @RequestParam(value = "page", defaultValue = "1") @Min(value = 1, message = "page must be greater than or equal to 1") int page,
-            @RequestParam(value = "per_page", defaultValue = "10000") @Min(value = 1, message = "per_page must be greater than or equal to 1") int per_page) {
-        return problemService.readAll(page - 1, per_page);
+            @RequestParam(value = "per_page", defaultValue = "10000") @Min(value = 1, message = "per_page must be greater than or equal to 1") int per_page,
+            @AuthenticationPrincipal User user) {
+        if (user == null)
+            return problemService.readAll(page - 1, per_page);
+        else {
+            return problemRepository.findAllProblemSummaryByUserId(user.getUserId(),
+                    PageRequest.of(page - 1, per_page));
+        }
     }
 
     @GetMapping("/{problemId}")
@@ -68,19 +86,38 @@ public class ProblemController {
     @PutMapping("/{problemId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void update(@PathVariable("problemId") Long problemId, @Valid @RequestBody ProblemForm problemForm,
-            HttpServletRequest request, HttpServletResponse response) {
-        Problem newProblem = new Problem(problemForm.getTitle(), problemForm.getSource(), problemForm.getDescription(),
+            HttpServletRequest request, HttpServletResponse response, @AuthenticationPrincipal User user) {
+
+        Problem problem = problemService.readOne(problemId);
+        if (problem == null) {
+            problem = new Problem();
+        } else if (problem.getCreateUser().getUserId() != user.getUserId() && !user.getRoles().contains("ROLE_GMH")) {
+            throw new CustomException("method not allowed", HttpStatus.BAD_REQUEST);
+        }
+
+        problem.update(problemForm.getTitle(), problemForm.getSource(), problemForm.getDescription(),
                 problemForm.getInput(), problemForm.getOutput(), problemForm.getSampleInput(),
                 problemForm.getSampleOutput(), problemForm.getHint(), problemForm.getTimeLimit(),
                 problemForm.getMemoryLimit());
-        newProblem.setProblemId(problemId);
-        problemService.update(newProblem);
+
+        problem.setProblemId(problemId);
+
+        problemService.update(problem);
     }
 
     @DeleteMapping("/{problemId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable("problemId") Long problemId, HttpServletRequest request,
-            HttpServletResponse response) {
+            HttpServletResponse response, @AuthenticationPrincipal User user) {
+
+        Problem problem = problemService.readOne(problemId);
+        if (problem == null)
+            return;
+
+        if (problem.getCreateUser().getUserId() != user.getUserId() && !user.getRoles().contains("ROLE_GMH")) {
+            throw new CustomException("method not allowed", HttpStatus.BAD_REQUEST);
+        }
+
         problemService.delete(problemId);
     }
 }
